refactor: restructure hub utils
