+++ gpustack/utils/hub.py
@@ -1,0 +1,147 @@
+import logging
+from typing import List
+from pathlib import Path
+import fnmatch
+from huggingface_hub import HfFileSystem
+from huggingface_hub.utils import validate_repo_id
+from modelscope.hub.api import HubApi
+
+from gpustack.config.config import get_global_config
+from gpustack.schemas.models import Model, SourceEnum
+
+logger = logging.getLogger(__name__)
+
+
+def match_hugging_face_files(repo_id: str, filename: str) -> List[str]:
+    validate_repo_id(repo_id)
+
+    hffs = HfFileSystem()
+
+    files = [
+        file["name"] if isinstance(file, dict) else file for file in hffs.ls(repo_id)
+    ]
+
+    file_list: List[str] = []
+    for file in files:
+        rel_path = Path(file).relative_to(repo_id)
+        file_list.append(str(rel_path))
+
+    matching_files = [file for file in file_list if fnmatch.fnmatch(file, filename)]  # type: ignore
+    matching_files = sorted(matching_files)
+    return matching_files
+
+
+def match_model_scope_file_paths(model_id: str, file_path: str) -> List[str]:
+    if '/' in file_path:
+        root, _ = file_path.rsplit('/', 1)
+    else:
+        root = None
+
+    api = HubApi()
+    files = api.get_model_files(model_id, root=root)
+
+    file_paths = [file["Path"] for file in files]
+    matching_paths = [p for p in file_paths if fnmatch.fnmatch(p, file_path)]
+    matching_paths = sorted(matching_paths)
+    return matching_paths
+
+
+def get_pretrained_config(model: Model):
+    """
+    Get the pretrained config of the model from Hugging Face or ModelScope.
+    Args:
+        model: Model to get the pretrained config for.
+    """
+
+    trust_remote_code = False
+    if model.backend_parameters and "--trust-remote-code" in model.backend_parameters:
+        trust_remote_code = True
+
+    global_config = get_global_config()
+    pretrained_config = None
+    if model.source == SourceEnum.HUGGING_FACE:
+        from transformers import AutoConfig
+
+        pretrained_config = AutoConfig.from_pretrained(
+            model.huggingface_repo_id,
+            token=global_config.huggingface_token,
+            trust_remote_code=trust_remote_code,
+        )
+    elif model.source == SourceEnum.MODEL_SCOPE:
+        from modelscope import AutoConfig
+
+        pretrained_config = AutoConfig.from_pretrained(
+            model.model_scope_model_id,
+            trust_remote_code=trust_remote_code,
+        )
+    else:
+        raise ValueError(f"Unsupported model source: {model.source}")
+
+    return pretrained_config
+
+
+# Simplified from vllm.config._get_and_verify_max_len
+# Keep in our codebase to avoid dependency on vllm's internal
+# APIs which may change unexpectedly.
+# https://github.com/vllm-project/vllm/blob/v0.6.2/vllm/config.py#L1668
+def get_max_model_len(pretrained_config) -> int:  # noqa: C901
+    """Get the model's maximum length."""
+    derived_max_model_len = float("inf")
+    possible_keys = [
+        # OPT
+        "max_position_embeddings",
+        # GPT-2
+        "n_positions",
+        # MPT
+        "max_seq_len",
+        # ChatGLM2
+        "seq_length",
+        # Command-R
+        "model_max_length",
+        # Others
+        "max_sequence_length",
+        "max_seq_length",
+        "seq_len",
+    ]
+    # Choose the smallest "max_length" from the possible keys.
+    max_len_key = None
+    for key in possible_keys:
+        max_len = getattr(pretrained_config, key, None)
+        if max_len is not None:
+            max_len_key = key if max_len < derived_max_model_len else max_len_key
+            derived_max_model_len = min(derived_max_model_len, max_len)
+
+    # If none of the keys were found in the config, use a default and
+    # log a warning.
+    if derived_max_model_len == float("inf"):
+        default_max_len = 2048
+        logger.warning(
+            "The model's config.json does not contain any of the following "
+            "keys to determine the original maximum length of the model: "
+            "%s. Assuming the model's maximum length is %d.",
+            possible_keys,
+            default_max_len,
+        )
+        derived_max_model_len = default_max_len
+
+    rope_scaling = getattr(pretrained_config, "rope_scaling", None)
+    if rope_scaling is not None:
+        if "type" in rope_scaling:
+            rope_type = rope_scaling["type"]
+        elif "rope_type" in rope_scaling:
+            rope_type = rope_scaling["rope_type"]
+        else:
+            raise ValueError("rope_scaling must have a 'type' or 'rope_type' key.")
+
+        # The correct one should be "longrope", kept "su" here
+        # to be backward compatible
+        if rope_type not in ("su", "longrope", "llama3"):
+            scaling_factor = 1
+            if "factor" in rope_scaling:
+                scaling_factor = rope_scaling["factor"]
+            if rope_type == "yarn":
+                derived_max_model_len = rope_scaling["original_max_position_embeddings"]
+            derived_max_model_len *= scaling_factor
+
+    logger.debug(f"Derived max model length: {derived_max_model_len}")
+    return int(derived_max_model_len)
