+++ gpustack/worker/backends/vllm.py
@@ -4,8 +4,9 @@ import subprocess
 import sys
 import sysconfig
 from typing import Optional
-from gpustack.schemas.models import ModelInstanceStateEnum, SourceEnum
+from gpustack.schemas.models import ModelInstanceStateEnum
 from gpustack.utils.command import find_parameter
+from gpustack.utils.hub import get_max_model_len, get_pretrained_config
 from gpustack.worker.backends.base import InferenceServer
 
 logger = logging.getLogger(__name__)
@@ -90,101 +91,10 @@ class VLLMServer(InferenceServer):
         Derive max model length from model config.
         Returns None if unavailable.
         """
-        trust_remote_code = False
-        if (
-            self._model.backend_parameters
-            and "--trust-remote-code" in self._model.backend_parameters
-        ):
-            trust_remote_code = True
-
-        pretrained_config = None
-        if self._model.source == SourceEnum.HUGGING_FACE:
-            from transformers import AutoConfig
-
-            pretrained_config = AutoConfig.from_pretrained(
-                self._model.huggingface_repo_id,
-                token=self._config.huggingface_token,
-                trust_remote_code=trust_remote_code,
-            )
-        elif self._model.source == SourceEnum.MODEL_SCOPE:
-            from modelscope import AutoConfig
-
-            pretrained_config = AutoConfig.from_pretrained(
-                self._model.model_scope_model_id,
-                trust_remote_code=trust_remote_code,
-            )
-        else:
-            # Should not reach here.
-            raise ValueError(f"Unsupported model source: {self._model.source}")
-
-        if pretrained_config:
+        try:
+            pretrained_config = get_pretrained_config(self._model)
             return get_max_model_len(pretrained_config)
+        except Exception as e:
+            logger.error(f"Failed to derive max model length: {e}")
 
         return None
-
-
-# Simplified from vllm.config._get_and_verify_max_len
-# Keep in our codebase to avoid dependency on vllm's internal
-# APIs which may change unexpectedly.
-# https://github.com/vllm-project/vllm/blob/v0.6.2/vllm/config.py#L1668
-def get_max_model_len(hf_config) -> int:  # noqa: C901
-    """Get the model's maximum length."""
-    derived_max_model_len = float("inf")
-    possible_keys = [
-        # OPT
-        "max_position_embeddings",
-        # GPT-2
-        "n_positions",
-        # MPT
-        "max_seq_len",
-        # ChatGLM2
-        "seq_length",
-        # Command-R
-        "model_max_length",
-        # Others
-        "max_sequence_length",
-        "max_seq_length",
-        "seq_len",
-    ]
-    # Choose the smallest "max_length" from the possible keys.
-    max_len_key = None
-    for key in possible_keys:
-        max_len = getattr(hf_config, key, None)
-        if max_len is not None:
-            max_len_key = key if max_len < derived_max_model_len else max_len_key
-            derived_max_model_len = min(derived_max_model_len, max_len)
-
-    # If none of the keys were found in the config, use a default and
-    # log a warning.
-    if derived_max_model_len == float("inf"):
-        default_max_len = 2048
-        logger.warning(
-            "The model's config.json does not contain any of the following "
-            "keys to determine the original maximum length of the model: "
-            "%s. Assuming the model's maximum length is %d.",
-            possible_keys,
-            default_max_len,
-        )
-        derived_max_model_len = default_max_len
-
-    rope_scaling = getattr(hf_config, "rope_scaling", None)
-    if rope_scaling is not None:
-        if "type" in rope_scaling:
-            rope_type = rope_scaling["type"]
-        elif "rope_type" in rope_scaling:
-            rope_type = rope_scaling["rope_type"]
-        else:
-            raise ValueError("rope_scaling must have a 'type' or 'rope_type' key.")
-
-        # The correct one should be "longrope", kept "su" here
-        # to be backward compatible
-        if rope_type not in ("su", "longrope", "llama3"):
-            scaling_factor = 1
-            if "factor" in rope_scaling:
-                scaling_factor = rope_scaling["factor"]
-            if rope_type == "yarn":
-                derived_max_model_len = rope_scaling["original_max_position_embeddings"]
-            derived_max_model_len *= scaling_factor
-
-    logger.debug(f"Derived max model length: {derived_max_model_len}")
-    return int(derived_max_model_len)
